pyproject.toml
[tool.poetry]
name = "mlstarterpack"
version = "0.1.0"
description = "A semi-random collection of utilities to kickstart ml projects"
authors = ["Anton Barkovsky <anton@swarmer.me>"]
license = "MIT license"
packages = [
    { include = "mlstarterpack", from = "src"},
]
include = ["LICENSE"]

[tool.poetry.dependencies]
python = ">=3.6,<4"

# Normal dependencies
numpy = "^1.16"
pandas = "^0.23.4"
pillow = "^5.4"
pyarrow = "^0.11.1"
pyyaml = "^5.3"

# Feature dependencies
tensorflow = { version = "~2.1.0", optional = true }

# Fake dependencies only used to build docs
Sphinx = { version = "^1.8", optional = true }
sphinxcontrib-apidoc = { version = "^0.3.0", optional = true }
sphinxcontrib-fulltoc = { version = "^1.2", optional = true }

[tool.poetry.extras]
docs = ["Sphinx", "sphinxcontrib-apidoc", "sphinxcontrib-fulltoc"]
tensorflow = ["tensorflow"]

[tool.poetry.dev-dependencies]
coverage = "^4.5"
ipython = "*"
mypy = "*"
pudb = "*"
pylint = "^2.2"
pytest = "^4.0"
pytest-runner = "^4.2"
tox = "^3.6"
watchdog = "^0.9.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
